+++ src/main/java/org/folio/marc/migrations/domain/entities/types/OperationStatusType.java
@@ -1,0 +1,14 @@
+package org.folio.marc.migrations.domain.entities.types;
+
+public enum OperationStatusType {
+
+  NEW,
+  DATA_MAPPING,
+  DATA_MAPPING_COMPLETED,
+  DATA_MAPPING_FAILED,
+  DATA_SAVING,
+  DATA_SAVING_COMPLETED,
+  DATA_SAVING_FAILED
+
+}
+